chore(ci): refactor linter script
